pyproject.toml
[project]
name = "gramps-webapi"
description = "A RESTful web API for the Gramps genealogical database."
authors = [
  { name = "Gramps Development Team" },
  { name = "David M. Straub", email = "straub@protonmail.com" }
]
license = "AGPL-3.0-or-later"
readme = "README.md"
dynamic = ["version"]
keywords = ["RESTful", "web API", "genealogy", "Gramps"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "gramps[all]>=6.0.4,<6.1.0",
    "PYGObject<=3.50.0",
    "orjson",
    "Click>=7.0",
    "Flask>=2.1.0",
    "Flask-Caching>=2.0.0",
    "Flask-Compress",
    "Flask-Cors",
    "Flask-JWT-Extended>=4.2.1,!=4.4.0,!=4.4.1",
    "Flask-Limiter>=2.9.0",
    "Flask-SQLAlchemy",
    "marshmallow>=3.13.0",
    "waitress",
    "webargs",
    "SQLAlchemy>=2.0.0",
    "pdf2image",
    "Pillow<12.0.0",
    "bleach[css]>=5.0.0",
    "jsonschema",
    "ffmpeg-python",
    "boto3",
    "alembic",
    "celery[redis]",
    "Unidecode",
    "pytesseract",
    "gramps-ql>=0.4.0",
    "object-ql>=0.1.3",
    "sifts>=1.1.0",
    "requests",
    "yclade>=0.5.0",
    "Authlib>=1.6.4",
    "geopy",
    "gramps-gedcom7",
]

[project.optional-dependencies]
ai = ["sentence-transformers>=4.1.0", "accelerate", "google-genai>=1.0.0"]
test = ["pytest", "PyYAML"]
all = ["sentence-transformers>=4.1.0", "accelerate", "google-genai>=1.0.0", "pytest", "PyYAML"]

[project.urls]
homepage = "https://github.com/gramps-project/gramps-web-api"
repository = "https://github.com/gramps-project/gramps-web-api"

[build-system]
requires = ["setuptools>=64.0", "setuptools_scm[toml]>=6.2", "wheel"]
build-backend = "setuptools.build_meta"

[tool.isort]
profile = "black"

[tool.setuptools.dynamic]
version = {attr = "gramps_webapi.__version__"}

[tool.setuptools.packages.find]
include = ["gramps_webapi", "gramps_webapi.*"]
